Moved old table files to legacy directory.
