+++ cpp/src/groupby/groupby_without_aggregation.cu
@@ -19,8 +19,8 @@
 #include <cudf/copying.hpp>
 #include <cudf/types.hpp>
 #include <string/nvcategory_util.hpp>
-#include <table/device_table.cuh>
-#include <table/device_table_row_operators.cuh>
+#include <table/legacy/device_table.cuh>
+#include <table/legacy/device_table_row_operators.cuh>
 #include <utilities/error_utils.hpp>
 
 #include <thrust/fill.h>
