+++ cpp/src/join/joining.h
@@ -22,7 +22,7 @@
 
 #include <cudf/functions.h>
 #include <cudf/types.h>
-#include <table/device_table.cuh>
+#include <table/legacy/device_table.cuh>
 
 #include "sort_join.cuh"
 #include "join_compute_api.h"
