+++ cpp/src/orderby/orderby.cu
@@ -22,8 +22,8 @@
 #include <utilities/cudf_utils.h>
 #include <utilities/error_utils.hpp>
 
-#include <table/device_table.cuh>
-#include <table/device_table_row_operators.cuh>
+#include <table/legacy/device_table.cuh>
+#include <table/legacy/device_table_row_operators.cuh>
 
 #include <rmm/thrust_rmm_allocator.h>
