+++ cpp/src/stream_compaction/stream_compaction.cu
@@ -16,8 +16,8 @@
 
 #include "copy_if.cuh"
 #include <cudf/table.hpp>
-#include "table/device_table.cuh"
-#include <table/device_table_row_operators.cuh>
+#include <table/legacy/device_table.cuh>
+#include <table/legacy/device_table_row_operators.cuh>
 #include <rmm/thrust_rmm_allocator.h>
 #include <string/nvcategory_util.hpp>
 #include <nvstrings/NVCategory.h>
