+++ cpp/src/table/legacy/table.cpp
