+++ cpp/tests/device_table/device_table_tests.cu
@@ -16,8 +16,8 @@
 
 #include <bitmask/legacy/bit_mask.cuh>
 #include <bitmask/legacy/bitmask_ops.hpp>
-#include <table/device_table.cuh>
-#include <table/device_table_row_operators.cuh>
+#include <table/legacy/device_table.cuh>
+#include <table/legacy/device_table_row_operators.cuh>
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 #include <tests/utilities/column_wrapper.cuh>
